requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

clip==1.0
matplotlib==3.5.3
medmnist==3.0.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
tensorboardX==2.6.1
timm==0.6.13
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.66.2
transformers==4.30.2
wandb==0.15.5
